refactor(web): case details overview file too big
